refactor: 수익 계산을 ProfitCalculator가 담당하도록 수정
